+++ webapp/controller/StorageBinsList.controller.js
@@ -5,18 +5,38 @@ sap.ui.define([
 	"sap/ui/model/Filter",
 	"sap/ui/model/FilterOperator",
 	"sap/m/MessageToast",
+	"sap/m/MessageBox",
 	"sap/ui/core/ValueState"
-], function (Controller, Formatter, JSONModel, Filter, FilterOperator, MessageToast, ValueState) {
+], function (Controller, Formatter, JSONModel, Filter, FilterOperator, MessageToast, MessageBox, ValueState) {
 	'use strict';
 
 	return Controller.extend('glw.controller.StorageBinsList', {
 		formatter: Formatter,
 		onInit: function () {
-
+			this.getOwnerComponent().getRouter().getRoute("storageBinsList").attachPatternMatched(this._updateStorageBinUsage.bind(this));
 		},
 
 		onNavBack: function () {
 			this.getOwnerComponent().onNavBack();
+		},
+
+		_updateStorageBinUsage: function () {
+			var mModels = this.getOwnerComponent().models;
+			Promise.all([mModels.container.loaded, mModels.storageBins.loaded]).then(function () {
+				var mStorageBins = {};
+				var aContainer = mModels.container.model.getObject("/rows");
+				for (var i = 0; i < aContainer.length; i++) {
+					mStorageBins[aContainer[i].value.storageBin] = true;
+				}
+
+				var aStorageBins = JSON.parse(JSON.stringify(mModels.storageBins.model.getObject("/rows")));
+				for (var j = 0; j < aStorageBins.length; j++) {
+					aStorageBins[j].value.used = mStorageBins[aStorageBins[j].value.id] || false;
+				}
+
+				mModels.storageBins.model.setProperty("/list", aStorageBins);
+				mModels.storageBins.model.refresh(true);
+			});
 		},
 
 		onDeleteStorageBinPress: function (oEvent) {
@@ -34,8 +54,26 @@ sap.ui.define([
 						width: "30rem",
 						duration: 2000
 					});
-				}};
-			oComponent.deleteDocument(oContext.getProperty("value")).then(fnHandler, fnHandler);
+				}
+			};
+
+			if (this._checkDeleteConditions(oContext.getProperty())) {
+				oComponent.deleteDocument(oContext.getProperty("value")).then(fnHandler, fnHandler);
+			} else {
+				var bCompact = !!this.getView().$().closest(".sapUiSizeCompact").length;
+				MessageBox.error(
+					"Der Lagerplatz wird noch von mindestens einem Behälter belegt",
+					{
+						styleClass: bCompact ? "sapUiSizeCompact" : ""
+					}
+				);
+			}
+		},
+
+		_checkDeleteConditions: function (oObject) {
+			return !this.getOwnerComponent().findEntity("container", "/rows", function (oContainer) {
+				return oObject.value.id === oContainer.value.storageBin;
+			});
 		},
 
 		onOpenAddStorageBinDialogPress: function () {
@@ -67,10 +105,10 @@ sap.ui.define([
 		onSaveNewStorageBinPress: function () {
 			var oDialog = this._getAddDialog();
 			var oModel = oDialog.getModel();
-			oModel.setProperty("/storageBinId/valueState", ValueState.None);
-			oModel.setProperty("/storageBinId/valueStateText", "");
+
 			var oObject = oModel.getObject("/");
-			if (oObject.storageBinId.value) {
+
+			if (this._checkSaveConditions(oObject, oModel)) {
 				var oComponent = this.getOwnerComponent();
 				oComponent.postDocument("storageBin", {
 					id: oObject.storageBinId.value
@@ -89,17 +127,36 @@ sap.ui.define([
 						});
 					}
 				});
+			}
+		},
 
-			} else {
+		_checkSaveConditions: function (oObject, oModel) {
+			oModel.setProperty("/storageBinId/valueState", ValueState.None);
+			oModel.setProperty("/storageBinId/valueStateText", "");
+			var bReturn = true;
+			// storageBinId must not be empty
+			if (!oObject.storageBinId.value) {
 				oModel.setProperty("/storageBinId/valueState", ValueState.Error);
 				oModel.setProperty("/storageBinId/valueStateText", "Bitte eine Lagerplatznummer eingeben");
+				bReturn = false;
 			}
+
+			// storageBinId must be unique
+			if (this.getOwnerComponent().findEntity("storageBins", "/rows", function (oStorageBin) {
+					return oObject.storageBinId.value === oStorageBin.value.id;
+				})) {
+				oModel.setProperty("/storageBinId/valueState", ValueState.Error);
+				oModel.setProperty("/storageBinId/valueStateText", "Die Lagerplatznummer existiert bereits");
+				bReturn = false;
+			}
+
+			return bReturn;
 		},
 
 		_getAddDialog: function () {
 			return this.byId("StorageBinAddDialog");
 		},
-		
+
 		onSearchStorageBin: function (oEvent) {
 			var sValue = oEvent.getParameter("newValue") || oEvent.getParameter("query");
 			var oTable = this._getTable();
